Add steps for getting started with Linux
Before this commit, there were no directions for us Linux users, and
now, there are! If my experiences are to be believed, there's some stuff
that needs figuring in order to get up and going with Python 3.7 _on
Linux_ so the docs I'm adding here are very careful to suggest how to
use Python 3.6 instead.

Along the way, I shuffled some headers a bit. Tried not to step too much
on tone and stuff, just wanted the new info to go somewhere that seemed
sensical.

ALSO, I just added `pip` to the requirements and removed the sections
that left users wondering what might be wrong if it wasn't there. It
was actually something I had to install right along with virtualenv.

ALSO ALSO the music _is_ awesome!!!

:heart: :heart: